feat: move all initialization functions and timed events to cron.ts
