+++ src/index.ts
@@ -44,7 +44,6 @@ import ProductController from './controller/product-controller';
 import ProductCategoryController from './controller/product-category-controller';
 import TransactionController from './controller/transaction-controller';
 import VoucherGroupController from './controller/voucher-group-controller';
-import BalanceService from './service/balance-service';
 import BalanceController from './controller/balance-controller';
 import RbacController from './controller/rbac-controller';
 import GewisAuthenticationController from './gewis/controller/gewis-authentication-controller';
@@ -65,8 +64,6 @@ import AuthenticationSecureController from './controller/authentication-secure-c
 import DebtorController from './controller/debtor-controller';
 import EventController from './controller/event-controller';
 import EventShiftController from './controller/event-shift-controller';
-import EventService from './service/event-service';
-import DefaultRoles from './rbac/default-roles';
 import WriteOffController from './controller/write-off-controller';
 import ServerSettingsStore from './server-settings/server-settings-store';
 import SellerPayoutController from './controller/seller-payout-controller';
@@ -93,25 +90,6 @@ export class Application {
     await this.connection.close();
     this.logger.info('Application stopped.');
   }
-}
-
-/**
- * Sets up the rbac and initializes the rbac controllers of the application.
- * @param application - The application on which to bind the middleware
- *                      and controller.
- */
-async function setupRbac(application: Application) {
-  // Synchronize SudoSOS system roles
-  await DefaultRoles.synchronize();
-
-  // Define rbac controller and bind.
-  const controller = new RbacController(
-    {
-      specification: application.specification,
-      roleManager: application.roleManager,
-    },
-  );
-  application.app.use('/v1/rbac', controller.getRouter());
 }
 
 /**
@@ -229,49 +207,20 @@ export default async function createApp(): Promise<Application> {
 
   application.roleManager = await new RoleManager().initialize();
 
-  application.app.use('/v1/stripe', new StripeWebhookController(
-    {
-      specification: application.specification,
-      roleManager: application.roleManager,
-    },
-  ).getRouter());
+  const options: BaseControllerOptions = {
+    specification: application.specification,
+    roleManager: application.roleManager,
+  };
+  application.app.use('/v1/stripe', new StripeWebhookController(options).getRouter());
 
   const tokenHandler = await createTokenHandler();
   // Setup token handler and authentication controller.
   await setupAuthentication(tokenHandler, application);
 
-  // Setup RBAC.
-  await setupRbac(application);
-
-  await new BalanceService().updateBalances({});
-  const syncBalances = cron.schedule('41 1 * * *', () => {
-    logger.debug('Syncing balances.');
-    new BalanceService().updateBalances({}).then(() => {
-      logger.debug('Synced balances.');
-    }).catch((error => {
-      logger.error('Could not sync balances.', error);
-    }));
-  });
-  const syncEventShiftAnswers = cron.schedule('39 2 * * *', () => {
-    logger.debug('Syncing event shift answers.');
-    EventService.syncAllEventShiftAnswers()
-      .then(() => logger.debug('Synced event shift answers.'))
-      .catch((error) => logger.error('Could not sync event shift answers.', error));
-  });
-  const sendEventPlanningReminders = cron.schedule('39 13 * * *', () => {
-    logger.debug('Send event planning reminder emails.');
-    EventService.sendEventPlanningReminders()
-      .then(() => logger.debug('Sent event planning reminder emails.'))
-      .catch((error) => logger.error('Could not send event planning reminder emails.', error));
-  });
-
-  application.tasks = [syncBalances, syncEventShiftAnswers, sendEventPlanningReminders];
+  application.tasks = [];
 
   // REMOVE LATER
-  const options: BaseControllerOptions = {
-    specification: application.specification,
-    roleManager: application.roleManager,
-  };
+  application.app.use('/v1/rbac', new RbacController(options).getRouter());
   application.app.use('/v1/authentication', new AuthenticationSecureController(options, tokenHandler).getRouter());
   application.app.use('/v1/balances', new BalanceController(options).getRouter());
   application.app.use('/v1/banners', new BannerController(options).getRouter());
